Move SparkRddBackend to spark_backend.py
